Generates data from two logistic regression trees.

This function generates data from two logistic regression trees: one with three apparent clusters (in terms of variance of the features) but a single logistic regression generating y | x, and one with a single apparent cluster but three different logistic regressions generating y | x given a categorical feature.

generateData(n = 100, scenario = "tree", visualize = F)

Arguments

n
The number of observations to draw.

scenario
The "no tree" scenario denotes the first scenario where there is a single logistic regression generating the data. The "tree" scenario generates data from the second data generating mechanism where there are three logistic regressions.

visualize
Whether (T) or not (F) to plot the generated data.

Examples

generateData(scenario = "tree")

#>     x1           x2          x3 y c
#> 1    2 -1.345371820 -0.42623658 1 1
#> 2    1  0.277273777 -1.16501291 0 1
#> 3    2  2.381767997 -0.89349075 1 1
#> 4    1 -1.695563511 -2.58896967 1 1
#> 5    1 -0.120377635 -1.35387672 0 1
#> 6    1  0.198630427 -0.83859287 1 1
#> 7    2  1.061932094 -0.36976885 0 1
#> 8    2 -0.359547036 -0.57537934 1 1
#> 9    2  2.976710905 -2.93865476 0 1
#> 10   2 -0.208180518 -1.26255759 0 1
#> 11   1  0.626476126  2.85532120 0 1
#> 12   2  1.472629166  0.93374090 0 1
#> 13   2 -0.589043033  2.98638065 1 1
#> 14   1 -1.559503465 -0.45822559 1 1
#> 15   1  2.673343440 -0.13626635 0 1
#> 16   2 -3.466603627 -0.27624218 1 1
#> 17   2  1.317906871 -1.79815165 0 1
#> 18   2  0.053710077 -1.25743072 1 1
#> 19   2  1.519243038  3.09945203 0 1
#> 20   2  0.648397732 -0.84337058 0 1
#> 21   1  3.136228808  1.91357327 0 1
#> 22   2 -1.799888729 -1.57135894 1 1
#> 23   1  2.384457300 -2.94881736 0 1
#> 24   2  2.931977463 -0.48445664 0 1
#> 25   1  0.007406665  1.40379379 1 1
#> 26   2 -3.677559582  1.70884470 1 1
#> 27   2  0.715855954  2.50742815 1 1
#> 28   2 -0.894837253 -2.68236331 1 1
#> 29   1  1.188304905  3.04686378 1 1
#> 30   1  0.434455065 -1.05471650 0 1
#> 31   2  1.108407906  0.23724714 0 1
#> 32   1  0.478440602  0.75935220 0 1
#> 33   2  1.614246531 -1.22999266 0 1
#> 34   4  0.325132294 -2.22328368 0 2
#> 35   4 -0.813738858  1.29169781 1 2
#> 36   3  1.336716968 -0.60677975 0 2
#> 37   3  0.893970866 -0.34110813 0 2
#> 38   3  2.453427002  1.40114426 0 2
#> 39   4  1.033913163 -0.69884382 0 2
#> 40   4 -1.921869945 -0.95631525 0 2
#> 41   4 -0.319716779  2.01556294 1 2
#> 42   3  2.844809808  0.27230308 0 2
#> 43   3  2.665294821  1.93876850 1 2
#> 44   4  0.849906747 -2.53207286 0 2
#> 45   3  0.023579181 -1.23149037 0 2
#> 46   4  0.574586008 -1.29321922 0 2
#> 47   4 -0.067705674  0.14826553 1 2
#> 48   3  0.051527861 -0.56348272 1 2
#> 49   3  0.253540161  1.08585623 1 2
#> 50   3  1.747540259 -2.69607303 0 2
#> 51   4 -0.066305996 -0.99561471 0 2
#> 52   4 -0.150552664 -0.93558973 1 2
#> 53   4 -0.425166853 -0.11944865 0 2
#> 54   3  2.311222471  0.65343714 0 2
#> 55   3  0.247753530  2.95635145 1 2
#> 56   4  1.961433540 -0.89513801 0 2
#> 57   3  1.932385317 -0.82876082 0 2
#> 58   4  0.889345411  1.04394995 0 2
#> 59   4 -0.424415526 -0.23349595 0 2
#> 60   4  1.883826038  2.02334729 0 2
#> 61   3  1.364758727 -1.60278461 1 2
#> 62   3 -1.392042158  1.59667612 1 2
#> 63   3  1.860271257 -1.96908265 0 2
#> 64   4  0.230196269  3.09554205 1 2
#> 65   4  1.577898868  0.19707452 1 2
#> 66   4 -1.131316819 -0.34753267 1 2
#> 67   6  0.404408973 -1.58807618 1 3
#> 68   6 -0.944978121  1.70685392 0 3
#> 69   6  1.302989741 -0.24039793 1 3
#> 70   6  2.590793276  0.94573969 1 3
#> 71   6  0.036281463  2.42543955 1 3
#> 72   5  0.552037765 -0.29024974 1 3
#> 73   5 -1.963806447 -2.41168776 1 3
#> 74   6  1.107932896 -1.32774619 0 3
#> 75   6  0.067309481 -0.64850145 1 3
#> 76   6 -1.572595800 -0.63243579 0 3
#> 77   6  2.591776636 -0.25574109 1 3
#> 78   5 -1.767899609  0.36871641 0 3
#> 79   6  0.979810067 -1.11862168 1 3
#> 80   6 -0.552849738 -0.41091620 0 3
#> 81   6 -0.899331965  2.73686841 0 3
#> 82   6  0.081907748  0.02135061 1 3
#> 83   6  2.561516138  0.28206412 1 3
#> 84   5 -1.641559463 -0.08120054 0 3
#> 85   6 -0.433922737  0.69241751 1 3
#> 86   6  3.311119441 -0.89515544 1 3
#> 87   6  0.778123510  1.89488735 1 3
#> 88   5 -2.107376904 -1.71799356 1 3
#> 89   5  3.022296720  1.62693533 1 3
#> 90   5 -1.782237506 -2.29349297 0 3
#> 91   5  0.285571211 -2.36061282 1 3
#> 92   5 -1.754603865 -0.17113895 0 3
#> 93   5 -0.057122339  0.16681377 0 3
#> 94   6  3.531306391  0.32120643 1 3
#> 95   5  2.090139389  0.83066374 1 3
#> 96   5 -0.840498538 -1.58927928 0 3
#> 97   6 -1.007189075 -2.41507399 0 3
#> 98   5  0.738657832 -0.50722659 1 3
#> 99   5 -1.769085774  0.30740663 1 3
#> 100  5 -1.598845550 -1.41115055 0 3